(MCOP-14) package uninstall action broken on windows
In the past the uninstall action did not work on Windows when used in
conjunction with Puppet 3.0.0 or greater. Reason being the Windows
provider would not initiate its package member until instances method
was called.

This commit sets up the provider's package member before using the
provider object allowing the agent actions to work on Windows again.